nvme-rc: Cleanup fc resource before module unloading
Before we unload the module we should cleanup the fc resources first,
basically reorder the shutdown sequence to be in reverse order of the
setup path.

Also unload the nvme-fcloop after usage.

While at it also update the rdma stop_soft_rdma before the module
unloading for the same reasoning.

Signed-off-by: Daniel Wagner <dwagner@suse.de>