nvme: Make test image size configurable
The reduce the overall runtime of the testsuite by making the default
size of the test image small. For verification jobs, the default can be
overwriten via the newly introduced nvme_img_size environment variable.

The default size of 350M is the minimum requirement of nvme/012 and
nvme/013. These tests rely on an xfs filesystem which requires this
minimum size.

Signed-off-by: Daniel Wagner <dwagner@suse.de>