nvme/rc: add reconnect-delay argument only for fabrics transports
The loop transport is also considered a fabric transport. But not
connect options are supported, in particular the reconnect-delay option.

nvme-cli can't figure this out as the kernel reports the option support
for all transports (union). Thus, we have to filter out this option for
this loop transport in blktests.

Signed-off-by: Daniel Wagner <dwagner@suse.de>